Reorganize src folder and add Api and Scene classes
