+++ src/Api.ts
@@ -1,0 +1,1 @@
+export class Api {}
